[l10n] Throw an error when message keys conflict
We changed collect-strings.js in the past to collect from multiple
directories, not just one. This introduces the potential for message
keys to conflict if:

  - Both directories have a file with the same relative path.
  - The two files with the same relative path both have a UIString
    object with the same property.

In this case the collect-strings.js will throw an error now and the
dev has to resolve this by renaming one (ideally the newly added)
of the conflicting property names.

R=alexrudenko@chromium.org

Bug: 1185727
Change-Id: I175acabda1499a185c3a48a7cf0ffbbcbb846843