[sources] Fix popover not working correctly for HTML and TSX.
The highlight range detection logic for JavaScript-like expressions
wasn't kicking in for HTML and TSX files, and therefore we only got
the fallback identifier-like token behavior, which works for quite
a lot of cases, but not reliably for everything.

Drive-by-refactoring: Reorganize the core highlight range detection
logic to be more unit test friendly and add some initial unit tests.

Bug: chromium:1383892, chromium:1361862, chromium:1383451
Change-Id: Ibad482ddfeded239e07d260bf7faeddfff629247