[CodeMirror] Update @codemirror/lang-html to 6.1.4.
This includes among other fixes, support for common TypeScript and JSX
script tag formats. In particular `<script lang=ts>` which is used a lot
in .vue files is now properly supported.

This also adds support for `<script type="text/jsx">`, which is used
quite commonly in the React world, especially for small demos.

Ref: https://github.com/codemirror/dev/issues/999
Bug: chromium:1383451, chromium:1384460
Change-Id: If9b3f4fa87c937a06885e5f0ad46fa8590e5f05e