[unittest] Select main target to activate in real-connection tests
To activate the test target, RealConnection tests assume that that there
is only one target. This is not true when debugging unittests. The main
test run and the debug test run happen in separate contexts that share
the same chrome instance, and the debug test run will generally be
repeated by developers. Instead of selecting the single target, this CL
searches targets for the test target by name.

Bug: none
Change-Id: Ida88002dc4486e539581f64c757094f289427bec