Shared Storage: Add DevTools per-page events view.
We add a per-page view that displays in a grid the
SharedStorageAccessedEvents that have occurred since the
last main frame navigation. A preview pane will show the
expandable JSON of the event when any of that event's cells
receive focus.

For the full initial Shared Storage DevTools UI prototype, see
https://crrev.com/c/devtools/devtools-frontend/+/3983292.

Bug:1372007,1218540
Change-Id: Ic35558652b878e05369f119f2e8109c569181d2a