[l10n] Don't ship en-US.json locale
Now that in-code UIStrings are as performant as the locales loaded
from JSON files, we don't actually need to ship and load en-US.json.

This CL removes it from the bundle and omits fetching en-US.json.

Note that the PRESUBMIT will still keep it populated since the
en-US.json is the source of truth for the translation pipeline
tooling. We'll update those scripts in a follow-up.

R=alexrudenko@chromium.org

Bug: 1185727
Change-Id: I254358abdc4c0e9f0db13a25b2ae0a4d25411f4b