[i18n] Make locale fetch paths configurable
This CL adds two new variables to i18n.gni that allow configuration
of the URL paths we use to fetch either remote or local translation
files.

Since the bundled locales (especially en-US.json) need to work in both
hosted and non-hosted mode, we use the URL relative to i18nImpl.ts.
This is a bit unfortunate but can't really be avoided.

R=alexrudenko@chromium.org

Bug: 1185727
Change-Id: I6cb40146d8f728320bb2f7d7425b54613a0ff6e7