[formatter] Correctly format unquoted HTML attribute values.
This is a pretty horrible work-around for a bug in the CodeMirror 5 HTML
tokenizer, which isn't easy to fix because it shares this code with the
XML parser[^1], and which is also not actively maintained anymore. The
real fix here is to migrate off of CodeMirror 5 also for formatting and
pretty printing and use CodeMirror 6 instead, but that's a bigger
project. For now we ducktape the problem by merging a '/' token
following a string token in the HTML formatter, which does the trick.

[^1]: https://github.com/codemirror/codemirror5/blob/742627a/mode/xml/xml.js#L137

Before: https://imgur.com/dD8RUbk.png
After: https://imgur.com/Z52f5LF.png
Bug: chromium:1382752, chromium:1385374
Change-Id: If2d63ab48a135acf92a83c84afdb5e6bfa694786
Fixed: chromium:1386026