[sources] Fix interaction of pretty-print and in-place formatting.
This makes the in-place editor action exclusive with the pretty-print
button (just like it's exclusive with the formatting button when the
"Automatically pretty print files" experiment is turned off). The code
and the layering here is very messy, so I went for minimal changes here
first, and will do follow-up refactorings that make this unit testable.

Fixed: chromium:1383453
Bug: chromium:1382752
Change-Id: Ic6d2d6d339a91b9221d976ef6bdf75c453e7ec69