[l10n] Change glob pattern when looking for all JS/TS files
This CL changes the glob pattern we use to find all the JS/TS files
we want to import UI strings for L10n for. Instead of searching in
the processes CWD and make the directory part of the glob, we only
search inside the directory itself and use a simple "**/*.{js,ts}"
glob.

This fixes the case where the passed in directory has one ore more
"../" in their prefix, which prevents the ignore list from working.

This also has the nice benefit that the paths we find are already
relative to "front_end", so we don't have to any wrangling
to get the "front_end-relative" path for the UI message keys.

R=kimanh@chromium.org

Bug: 1185727
Change-Id: I844756b6a6143afc2bf3af35fe785f781e12a382