[l10n] Move locale targets into 'locales' directory
In an upcoming CL we want to reference the generated en-US.json and
en-XL.json files with the minify_json action. This is impossible the
way it's currently implemented as minify_json makes some assumptions
about it's input files.

To keep more with GN standards, we put the call to `minify_json` into
the 'locales' directory. This way, the rule doesn't have to do
something complicated but just take all the source files and put them
into `$target_gen_dir`.

R=kimanh@chromium.org

Bug: 1185727
Change-Id: I1748cbabb0dc76a16b445555132226429c13d7de