[e2e] Re-enable linear memory inspector wasm test
This changes the test to not explicitly pause the second worker, but
instead to just continue the execution and automatically run into
the second worker's debugger statement. Afterwards, we can inspect
the memory buffers.

This doesn't change the fact that we still look at the same memory
buffers (which is tested in this test).

Fixed: 1169143
Change-Id: Ia9030c4cf7265e0ad63d18281f0167a2f934271b