Focus datagrid cell synchronously on key press
DataGrid cells that have a `title` attribute show a tooltip on mouse
hover, which is expected platform behavior. Users that navigate the
DataGrid using the keyboard should also be able to reveal the tooltip.
Although cells can be focused by keyboard today, the platform does
not show the element's tooltip when this happens.

This appears to be because #focusTableCellInDOM uses the render
coordinator to schedule the call to focus() asynchronously. Calling
focus() invokes the platform tooltip only if it happens during a
trusted event (i.e. the keydown event). The fix is to remove the
coordinator.write usage and call focus() synchronously.

Bug: 1392881
Change-Id: I8afd88ee9b5f4dd17d1c3c4de1521563b8e7bbbe