[l10n] Properly rebuild en-US.json/en-XL.json when front_end changes
We need to run the ":collect_strings" action any time something
in "front_end" changes. This means ":collect_strings" has a dependency
to all the JS/TS files in "front_end".

To prevent a circular dependency, we no longer deps the locale file
in together with the i18n module, but instead depend on it
globally. This has the added benefit that we no longer re-build
half of DevTools just because one UI string was added.

R=alexrudenko@chromium.org

Bug: 1185727
Change-Id: I43352172bb91bbfbed8f19923743cff098ce3ba4