[TextEditor] Fix highlight line effect.
This simplifies the "highlighted line" effect in the TextEditor and
makes it always restart the CSS animation by explicitly clearing the
.cm-highlightedLine class before setting it again (potentially on the
same line). This way the animation always shows when you clear to set a
breakpoint or you jump to a file, even if that line was previously
highlighted already.

Also don't remove the .cm-highlightedLine class from the DOM with a
setTimeout() to avoid some random flakiness on our E2E tests. This class
isn't used anywhere else and the CSS animation only restarts when the
class is added, so there's no harm in leaving the class there.

Before: https://imgur.com/wAHaGJH.mp4
After: https://imgur.com/wplUvKg.mp4
Fixed: chromium:1382400
Change-Id: Ic6f80d38c09ad7e7bd185e0763b3ef016029d153