Add CSS aliases to autocomplete
This CL adds alias properties to CSS property name's autocomplete list,
together with their alias-for properties. For example, if `word-wrap`
is typed, we will now prompt both `word-wrap`, an alias property for
`overflow-wrap`, and `overflow-wrap`. It also shows an explanation
in the format of:
word-wrap    = overflow-wrap
in the autocomplete prompt.

Fixed: 998025
Change-Id: Ic4d5f68e4c017e433d04ee57b5f45d9a7e6782b2