[sources] Apply space setting immediately for pretty printing.
When the "Automatically pretty print files" experiment is turned on,
previously we wouldn't immediately re-pretty-print when the space
setting changed.

This CL takes a pragmatic approach and simply toggles the pretty
print state of the SourceFrame when the global setting changes.
This isn't the most beautiful solution, but the most pragmatic
at this point to resolve the last open blocker for the experiment
in M-110.

Bug: chromium:1382752
Change-Id: I136cacb58f08b04fc7594c97f50713c383841207
Fixed: chromium:1382393