[breakpoints] Prevent leaks of UISourceCode instances
Breakpoints keep references to "home" UI source codes to be able to set
set breakpoints on them (in addition, breakpoints hold references to the
visible locations where the breakpoints are bound). In some cases, the
visible UI source code can be different from the "home" UI source codes
(e.g., with source maps or with pretty printing). If that happens and
the home UI source code is removed (e.g., on reload), the breakpoint can
keep refering to a stale UI source code.

This patch adds accounting of breakpoints by their "home" UI source
code, so that the UI source code instances can be reliably removed.

Bug: chromium:1355758
Change-Id: Ic5ac55031c74f8280984bccc37c2284f7c7fb8a7