[l10n] Don't rebuild half of DevTools when en-US.json changes
This CL changes who depends on the "locales" target. Currently, the
i18n ts_library has the locale files as a public_deps. This means,
everytime en-US.json changes, ninja thinks that the i18n library
changed and rebuils everything that relies on core/i18n:i18n (which
is nearly all of DevTools).

It's enough that the top-level "front_end" target pulls in the
locale files.

R=alexrudenko@chromium.org

Bug: 1185727
Change-Id: I1d04634e4fb4db1d7716227175c02e1de1109a6f