[l10n] Only bake strings currently present in the source
This CL modifies the "npm run bake-strings" command to only include
UI strings in the baked *.json files that are also present in the
TS files.

It does this by first reading the en-US.ctc.json file and memorizing
all the keys present there. While converting the various
<locale>.ctc.json to <locale>.json we only write those messages
that were also present in en-US.ctc.json.

R=alexrudenko@chromium.org

Bug: 1185727
Change-Id: I2b9c099a12a49b53b8613c6a962a776db7d609b6