Fix console context selector showing white text
`toolbar-shadow` was establishing a stacking context when its z-index
was set. We were using `z-index: -1` to put a background with `::before`
pseudo element for the console context selector (by setting the background
of button transparent.)

After removing `z-index: 12` in crrev/c/4041405 that stacking context is
gone and caused this behavior.

This CL brings back the stacking index for the toolbar and sets
`toolbar-shadow`s z-index: 0 to preserve the behavior in crrev/c/4041405

Fixed: 1394045
Change-Id: I8e13c07eb5954384bbbe7dee7b396ee86afae497
Bug: chromium:1379758