[console] Introduce 'noJavaScriptCompletion=true' query parameter.
This introduces a new noJavaScriptCompletion query parameter that
when set to 'true' will disable the dynamic JavaScript engine
based autocompletion for the Console panel. This is useful, for
example in the context of ruby/debug, to suppress the unhelpful
JavaScript suggestions for the autocomplete when working in a
different programming language.

Fixed: chromium:1393972
Also-by: Naoto Ono <onoto1998@gmail.com>
Change-Id: I3e26e748a3fbd127cf9625ed3c023a46f97c8935