[breakpoints] Improved handling of breakpoints in inline scripts
If a user adds a breakpoint in an inline script before the embedding
document is loaded, the breakpoint will not be persisted under the
correct URL, so the breakpoint does not persist across reloads.

This patch makes sure that we use the emedding document's URL for
breakpoints in inline scripts. This unfortunately requires some
tricky translation between the inline script's and the embedding
document's line-column coordinates.

Bug: chromium:1355758
Change-Id: I903d3dd5656fdb67da4ecf25c89bbf4eb7983b1c