Shared Storage: Add DevTools tree elements.
This CL adds the TreeElement classes to hold both the per-page and
per-origin views.

We also fix a bug in the `SharedStorageEventsView` so that we are
actually getting the `mainFrameTarget` in order to get the main frame's
`ResourceTreeModel` as intended.

For the full initial Shared Storage DevTools UI prototype, see
https://crrev.com/c/devtools/devtools-frontend/+/3983292.

Bug:1372007,1218540
Change-Id: If765c317124578e32bd77832bf8f314d6f3ad3ac