Shared Storage: Fix minor view bugs.
In this CL we fix an issue where an entry with an null value
can fail to preview in the `SharedStorageItemsView`.

We also fix a bug in the `SharedStorageEventsView` so that we are
actually getting the `mainFrameTarget` in order to get the main frame's
`ResourceTreeModel` as intended.

For the full initial Shared Storage DevTools UI prototype, see
https://crrev.com/c/devtools/devtools-frontend/+/3983292.

Bug:1372007,1218540
Change-Id: I87fee438c8605e8ceadef7706e422c4cde5d563a