Fix line breaking behaviour with HeaderSectionRow
Remove the flex attribute for `header-name` and `header-value` which
was introduced in https://crrev.com/c/4061714. Instead set the `host`
attribute for the `EditableSpan` component to `inline`, in order to
ensure that the `EditableSpan` stays on the same line with its
surrounding elements.

Screenshots: https://imgur.com/a/J0Chd8P

Bug: 1297533
Change-Id: I70b5958d8855e1dd471511c5c76bcfb529766234