Add ESLint rule for commented out import statements
This is something I've been caught out with a few times when moving code
round, and thought it would be a good ESLint rule. I've enabled the rule
as part of this CL as we currently have no instances of it, and so
there's no huge diff here to make the code pass the new check.

Bug: none
Change-Id: I72c6aad3e4258fbda280e062f98823a068c6bfc8