Break up the Color class into per-space types
Implementing color conversions on a monolithic Color type becomes very messy. As a preparation, this CL takes apart the Color type, breaking it into one class per Color type with a common interface. Lots of places in the codebase make use of the Color type for purposes beyond displaying css styles and most of these assume colors in a legacy format. This CL introduces an explicit conversion function to the legacy format that all users need to call, which gives us a "marker" for all places assuming the legacy format.

Bug: 1395767, 1395782
Change-Id: If986eef0a9e168d2d08946e83ac21b0f7b093674