Header overrides editor: show warning for disallowed header names
Header names should only contain alphanumeric characters, '-', or '_'.
With this CL a warning is shown to the user when they enter characters
which are not allowed.

However, there is no enforcement of this rule. It is still possible to
enter disallowed characters and they will be persisted to the
'.headers' file. The corresponding request override will fail and the
response without any overrides will be used by Chrome.

Fixed: 1383088
Change-Id: I273716ebb228c6f3c3b72a6e6b133bb0c9b70b08