[sources] Origins for source-mapped stylesheets.
This generalizes the `ScriptOriginPlugin` to an `ResourceOriginPlugin`,
which not only provides origin links for scripts, but also for style
sheets.

As part of this, we migrate the source-frame-toolbar-items.js layout
test to an E2E test (in `sources/eval_test.ts`) and add more unit and
E2E test coverage for the refactored plugin.

Before: https://imgur.com/5RtouGK.png
After: https://imgur.com/C07dBdk.png
Fixed: chromium:1392106
Bug: chromium:1149734
Change-Id: I363f00aa0d9373682e9fcb695ff2a39c7fc69232