Settings-shortcuts: fix behaviour when clicking on rows
Previously after clicking on a row, the pencil icon was shown, but the
background color of the row stayed unchanged. This CL ensures that the
background color changes, so that it's clear that a row has been
selected.

Before: https://i.imgur.com/fJ9COqm.mp4
After: https://i.imgur.com/SHBaNUM.mp4

Bug: none
Change-Id: I4e3bf9b475b900ddc473ca61b032e9a47e5feca3