Make InspectorFrontendHost "worker-proof"
If a worker happens to import code from SDK, this file will also end up
in the worker context. This is mostly OK - most of the `window`
references are in called functions, but there is a reference to `window`
in the top level which errors.

Similarly, the `constructor()` runs and therefore we need to add the
check for `document` to avoid it.

In an ideal world no worker should ever load this file, even by
accident, but if they do, let's make sure that it doesn't immediately
error.

Bug: none
Change-Id: I5d7540927c8c1eac3c64e623e254f9f5c21c94bb