[Perf Panel Node] Able to download cpuprofile in Node mode
In Node mode, users are more willing to have .cpuprofile for later analysis.

This CL seperate the downloading logic for PerfPanel for Node, so for Node, we will directly download the stringfied CpuProfile (same as JS Profiler).
For web version, everything is the same, and the js samples we recorded in web version will be converted to TraceEvents and will be downloaded with other TraceEvent.

This CL also leave the possibility that we can skip the generating download-able file, for example when you load a file, you don't need to re-download it.

Screenshot: https://imgur.com/a/Ub1deMO
Bug: 1392446
Change-Id: Ib16fcf1258f9fd4838c15fb6de28f01d8f3da868