[TextEditor] Make autocompletion consistent.
This change generalizes the 'textEditorAutocompletion' setting to apply
consistently to all instances of TextEditor that can use autocompletion
(in particular the Console and the Elements panel). It also changes the
way we interpret the 'textEditorAutocompletion' setting such that when
it's turned off, autocompletion is still available but requires explicit
user interaction to trigger (Ctrl+Space) rather than showing up when
typing regularly.

Fixed: chromium:1399436
Change-Id: I5ef805a1df1f2f2631e07ab8b46578216c03b951