Remove language plugin test and replace breakpoint location clash test
This CL removes the language plugin test breakpoint restoration test,
for which we already have a replacement that makes use of the new mock CDP connection, see https://crrev.com/c/4097064.

Removing the language plugin test caused failures in another,
actually unrelated test, and indicated that the tests are leaking.
This CL removes the failing old location clash test, too, and
replaces it with a simpler and cleaner implementation with the
new mock connection.

Bug: 1396298, 1395337
Change-Id: I531457fc679980d5f2a5c835fcd198fc92f76b86