[A11y] Announce number of selectors shown after filtering styles
Adds an alert to announce the number of results found when using the
Filter box in the Styles pane. To avoid spamming the screen reader,
the alert is only executed once the keyboard had been idle for 500ms
(adjustable via the FILTER_IDLE_TIME constant). There are no UI changes
necessary for this fix.

Bug: 1379450
Change-Id: Ibc092262da4c5823b0325a056065fd140cce7e2f