check button value for shift + f10 keyboard event
An additional check of PoinerEvent.button value is added in this change to tell whether the event is from mouse click or "shift + f10" keyboard press. Note that a "shift + f10" keyboard press will invoke a context menu similarly as a mouse right click. The pointerType value is mouse in both the mouse right click and "shift + f10" keyboard press.

Bug: 1394898
Change-Id: Icd5d853f27d7b51dd799ed2dc3479a6341ba92e0