[console] Differentiate conservative completion visually.
This changes the visualization for conservative completion to not show
the current match as selected, but draw a dotted border around it. Only
once the user starts interacting with the autocompletion, we change to
regular selection visualization. This is inspired by the UI found in
Visual Studio.

Before: https://imgur.com/IuIhRuY.png
After: https://imgur.com/1lbIVrq.png
Bug: chromium:1276960
Change-Id: I9b1bf0c9e93bc74cdf07c0f02760851224ce3ceb