[console] Fix 'Store function as global variable' for bound functions.
For bound functions we try to figure out the target function and display
that when logging to the Console, which is much more useful than showing
the bound function (which is `function() { [native code] }`). This makes
sense generally, but due to the way it was implemented, it made it
impossible to store the bound function in a global variable via the
context menu (rather we would store the target function).

This change adopts the Firefox DevTools approach of showing the target
function (with an info icon that makes it clear that we looked through a
bound function here), but have the context menu operate on the bound
function. When jumping to the function definition however, we again look
through the bound function and jump to the target function.

Fixed: chromium:1330921
Change-Id: I8f35a1ff3022d4bb533b78112a2794550f2e748f