[Perf] Restrict events used for trace bounds based on phases
Before this CL, the max time of the trace window was determined by the
trace event with the largest timestamp across all collected events. This
is problematic because async events might yield incorrect values that
potentially spoil the scale of the timeline.

Instead, just as we do for the lower limit of the trace window, this CL
constrains the events considered for tracking the window bounds to those
with valid sync phases.

Fixed: 1400699
Change-Id: I5ed907906a2f0c2c087118b3ee63d62fc6ac8a9d