[sources] Fix missing ignore list infobar.
The ignore list infobar was not showing properly when (re)opening files
in the 'Sources' panel that were on the ignore list. This regressed as
part of the CM6 migration, because the editor is initialized after the
plugin is constructed now. We already handle this case properly for the
missing debug infobar, and the ignore list infobar needs to follow the
same approach here.

Fixed: chromium:1401618
Change-Id: I8839216070cc05c5dab5d394dddd430a07060607