Show info icon prompting user to reload upon header edits
When the user edits a header override, after making changes, the state
shown in the network panel's header view no longer reflects the state
of the headers when the response was received. This CL adds an info
icon which prompts the user to refresh so that their changes take
effect.

Video: https://i.imgur.com/gUtbIaj.mp4

Bug: 1383094
Change-Id: I1609d6cdf9b8173ebd16e82818b421fed4c399f0