Handle non-breaking space in header overrides
When comparing header values (or header names), until now we checked
for exact equality. This means that strings differing only in non-
breaking space versus regular space would not be considered equal.

When the user edits a contenteditable span/div, the result can contain
non-breaking spaces (which is not obvious to the user). It would be
difficult to replace those NBSPs during the editing process
(triggering unwanted re-renders, etc.). Therefore this CLs ensures
that during various header comparisons, NBSPs and regular spaces are
considered as the same character.

Fixed: 1401713
Change-Id: Iad5be08f272615599caf43e18350ed3dc4525fc4