Show deleted header overrides
When the user deletes a header override, instead of immediately
disappearing from the UI, render it with a red background color and
strikethrough font. Also show the info icon which instructs the user
to refresh the request for the changes to take effect.

Screenshot: https://i.imgur.com/Xf8eshy.png

Fixed: 1383094
Change-Id: I277610b60ec2806b08ec87701a0db6c13671b2aa