Update header overrides upon change to file
When a '.headers' file is being changed via an external editor, so far
the editor of the respective '.headers' file in the sources panel
would reflect these changes automatically. However, when refreshing
the request, the updated header overrides would not be used. Instead
the network panel headers view would still show the header overrides
in its state before the external edit. This CL fixes this and ensures
that whenever a '.headers' file is edited, this also affects the
interception patterns DevTools uses for header overrides.

Fixed: 1402230
Change-Id: I7d49c8633ed10964289f424bfaf4180facbe424e