Add the new "Layerize" devtools timeline event (front-end side)
Layerize is the step of the rendering pipeline where we decide which
layers to create [1]. It is the second most expensive step in the
pipeline [2] which is why it is important to not have it show a blank
space on the devtools timeline. This event is also actionable using the
"Layers" panel, or other knowledge of what "creates a layer".

This patch implements the frontend code and will be followed up with a
chromium-side patch to emit this new event.

[1] https://developer.chrome.com/articles/renderingng-architecture/
[2] https://uma.googleplex.com/p/chrome/timeline_v2?sid=d57f93f88aad2b81484351dca9cd716f

Bug: 1191032
Change-Id: I1f9f9173b23b2f255edcf05e89c28a16230a7de6