Shared Storage: Add reset budget button.
We add a button to the `SharedStorageMetdataView` to reset the entropy
budget. This will make it easier for developers who are testing; if
they exhaust the budget, they won't have to wait for it to be reset
or use workarounds like clearing browsing data or logging in from a
different profile.

Bug:1372007,1218540
Change-Id: I198c3957dbb7877f3958d63983b19ce921831c0b