Change hit testing for Elements tree items.
Previously, hovering or clicking anywhere within the same row as an
DOM tree element would highlight or select that tree element. However,
the implementation to achieve this is error-prone.

Now, only clicking on the visible content of the DOM tree element would
select that tree element. Clicking/hovering in the indent before that
tree element would, depending on the indent position, select/highlight
the parent on that level of indent. This makes the implementation
simpler/cleaner, and the new behavior is rather useful.

Fixed: 1376377
Change-Id: I427771d29cde0e4bc2cdc561bab5224e5cb92e91